Rename package earfuck to eflang inside ef-core
