+++ ef-core/build.gradle
@@ -4,5 +4,5 @@ dependencies {
     compile name: 'jfugue-4.0.3'
 }
 
-mainClassName = 'earfuck.Main'
+mainClassName = 'eflang.Main'
