+++ ef-core/src/main/java/eflang/SystemIoManager.java
@@ -1,4 +1,4 @@
-package earfuck;
+package eflang;
 
 import java.util.Scanner;
